PicoChat-pi-sdk/monitor/requirements.txt
pySerial
pyside6